Fix for JSON API 404ing on 'new.reddit' subdomain
Currently, for whatever reason, the 'new.reddit.com' subdomain 404s if you use any of the JSON API paths. 

For example:  
https://www.reddit.com/r/todayilearned/about.json (200 OK)  
https://old.reddit.com/r/todayilearned/about.json (200 OK)  
https://np.reddit.com/r/todayilearned/about.json (200 OK)  
https://new.reddit.com/r/todayilearned/about.json (404 NOT FOUND)  

It's probably a rare issue that would only affect a few people, but I figured I could attempt to solve it globally by swapping out the `new.` paths with `www.` in the `ajax()` function.