Makes video timeline control much easier to use
I frequently misclicked when trying to move the timeline position on desktop, as it's only a 2px area that you have to get over. So I adjusted it so that the clickable space is larger, but still appears as a 2px bar. 

I also made the timeline slightly more visible.